feat(build)!: pretty much complete overhaul
BREAKING CHANGE: This is no longer a drop in replacement for the old
4.21 Vis.js.

This brings this project inline with our other packages (except for TS
support, there is none).

I also removed the notice about this being a drop in replacement for the
old Vis and broke backwards compatibility by simply reexporting up to
date packages. I don't have the time to manage backwards compatibility
here and I don't even think it's worth it at all.

All the issues that have accumulated over the past 7 months (since the
last release) are resolved here.