fix(types): add missing dummy declarations
It's not currently possible to mix the types from the packages. I don't
think it's worth the effort. If someone wants to use Vis in TS they can
easily use the packages directly. This is more of a script tag thing and
maybe also a kind of smoke test for issues with combining the different
projects.