feat(build)!: use a compromise solution
BREAKING CHANGE: This basically reverts the most breaking parts of the
previosly commited overhaul.

Use a compromise solution between backwards compatibility and easy of
maintenance. The old exports are restored and so is the file listing in
the package. However the libraries themselves may still introduce
breaking changes in their features etc.