feat(build)!: pretty much complete overhaul (#88)
* feat(build)!: pretty much complete overhaul

BREAKING CHANGE: This is no longer a drop in replacement for the old
4.21 Vis.js.

This brings this project inline with our other packages (except for TS
support, there is none).

I also removed the notice about this being a drop in replacement for the
old Vis and broke backwards compatibility by simply reexporting up to
date packages. I don't have the time to manage backwards compatibility
here and I don't even think it's worth it at all.

All the issues that have accumulated over the past 7 months (since the
last release) are resolved here.

* feat(build)!: use a compromise solution

BREAKING CHANGE: This basically reverts the most breaking parts of the
previosly commited overhaul.

Use a compromise solution between backwards compatibility and easy of
maintenance. The old exports are restored and so is the file listing in
the package. However the libraries themselves may still introduce
breaking changes in their features etc.

* chore(readmy): update copyright

It's 2020 already.

* chore(readme): use latest in usage

This way people will hopefully pick the latest version instead of
blindly copypasting 1.0.0.

* fix(build): also restore DOMutil